set sequence as uint32_t type.

+++ netc/Timer.h
@@ -23,7 +23,7 @@ private:
     const TimerCallback callback;   //定时器回调函数
     Timestamp expiration;           //下一次超时时刻
     const double interval;          //超时间隔，如果是一次性定时器，则设为 0
-    const int64_t sequence;         //定时器序号，从 1 开始
+    const uint32_t sequence;        //定时器序号，从 1 开始
     static atomic_uint32_t num_created;  //定时器计数，当前已经创建的定时器个数
 public:
     /**
